Move some project-env-info tests to "integration" style
