+++ project_test.go
@@ -16,7 +16,6 @@ import (
 	"testing"
 
 	"github.com/tsuru/gnuflag"
-	"github.com/tsuru/tsuru-client/tsuru/client"
 	"github.com/tsuru/tsuru/cmd"
 )
 
@@ -374,8 +373,8 @@ func TestProjectUpdateFailToCreateNewApps(t *testing.T) {
 		payload: []byte(listOfApps),
 	})
 	appRespMap := map[string][]byte{
-		"proj3-dev":  []byte(appInfo2),
-		"proj3-prod": []byte(appInfo3),
+		"proj3-dev":  []byte(appInfo1),
+		"proj3-prod": []byte(appInfo2),
 	}
 	for appName, payload := range appRespMap {
 		server.prepareResponse(preparedResponse{
@@ -427,8 +426,8 @@ func TestProjectUpdateFailToSetCNames(t *testing.T) {
 		payload: []byte(listOfApps),
 	})
 	appRespMap := map[string][]byte{
-		"proj3-dev":  []byte(appInfo2),
-		"proj3-prod": []byte(appInfo3),
+		"proj3-dev":  []byte(appInfo1),
+		"proj3-prod": []byte(appInfo2),
 	}
 	for appName, payload := range appRespMap {
 		server.prepareResponse(preparedResponse{
@@ -490,8 +489,8 @@ func TestProjectUpdateFailToUpdate(t *testing.T) {
 		payload: []byte(listOfApps),
 	})
 	appRespMap := map[string][]byte{
-		"proj3-dev":  []byte(appInfo2),
-		"proj3-prod": []byte(appInfo3),
+		"proj3-dev":  []byte(appInfo1),
+		"proj3-prod": []byte(appInfo2),
 	}
 	for appName, payload := range appRespMap {
 		server.prepareResponse(preparedResponse{
@@ -563,8 +562,8 @@ func TestProjectUpdateInvalidNewEnv(t *testing.T) {
 		payload: []byte(listOfApps),
 	})
 	appRespMap := map[string][]byte{
-		"proj3-dev":  []byte(appInfo2),
-		"proj3-prod": []byte(appInfo3),
+		"proj3-dev":  []byte(appInfo1),
+		"proj3-prod": []byte(appInfo2),
 	}
 	for appName, payload := range appRespMap {
 		server.prepareResponse(preparedResponse{
@@ -607,8 +606,8 @@ func TestProjectUpdateDuplicateEnv(t *testing.T) {
 		payload: []byte(listOfApps),
 	})
 	appRespMap := map[string][]byte{
-		"proj3-dev":  []byte(appInfo2),
-		"proj3-prod": []byte(appInfo3),
+		"proj3-dev":  []byte(appInfo1),
+		"proj3-prod": []byte(appInfo2),
 	}
 	for appName, payload := range appRespMap {
 		server.prepareResponse(preparedResponse{
@@ -655,8 +654,8 @@ func TestProjectUpdateInvalidRemoveEnvs(t *testing.T) {
 		payload: []byte(listOfApps),
 	})
 	appRespMap := map[string][]byte{
-		"proj3-dev":  []byte(appInfo2),
-		"proj3-prod": []byte(appInfo3),
+		"proj3-dev":  []byte(appInfo1),
+		"proj3-prod": []byte(appInfo2),
 	}
 	for appName, payload := range appRespMap {
 		server.prepareResponse(preparedResponse{
@@ -840,58 +839,6 @@ func TestProjectInfoMissingName(t *testing.T) {
 	expectedMsg := "please provide the name of the project"
 	if err.Error() != expectedMsg {
 		t.Errorf("wrong error message\nwant %q\ngot  %q", expectedMsg, err.Error())
-	}
-}
-
-func TestProjectEnvInfo(t *testing.T) {
-	fakeServer := newFakeServer(t)
-	defer fakeServer.stop()
-	fakeServer.prepareResponse(preparedResponse{
-		method:  "GET",
-		path:    "/apps/proj1-prod",
-		code:    http.StatusOK,
-		payload: []byte(appInfo1),
-	})
-	fakeServer.prepareResponse(preparedResponse{
-		method:  "GET",
-		path:    "/apps/proj1-prod/quota",
-		code:    http.StatusOK,
-		payload: []byte(`{"Limit":10,"InUse":1}`),
-	})
-	fakeServer.prepareResponse(preparedResponse{
-		method: "GET",
-		path:   "/services/instances?app=proj1-prod",
-		code:   http.StatusNoContent,
-	})
-	cleanup, err := setupFakeTarget(fakeServer.url())
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer cleanup()
-	var stdout, stderr bytes.Buffer
-	ctx := cmd.Context{Stdout: &stdout, Stderr: &stderr}
-	cli := cmd.NewClient(http.DefaultClient, &ctx, &cmd.Manager{})
-	var buf bytes.Buffer
-	var appInfoCmd client.AppInfo
-	err = appInfoCmd.Flags().Parse(true, []string{"-a", "proj1-prod"})
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = appInfoCmd.Run(&cmd.Context{Stdout: &buf}, cli)
-	if err != nil {
-		t.Fatal(err)
-	}
-	var c projectEnvInfo
-	err = c.Flags().Parse(true, []string{"-n", "proj1", "-e", "prod"})
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = c.Run(&ctx, cli)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if stdout.String() != buf.String() {
-		t.Errorf("Wrong output\nWant:\n%s\nGot:\n%s", &buf, &stdout)
 	}
 }
