+++ stub_test.go
@@ -192,70 +192,6 @@ const deployments = `[
 ]`
 
 const appInfo1 = `{
-  "cname": ["proj1.example.com"],
-  "deploys": 35,
-  "description": "my nice project",
-  "ip": "proj1-prod.tsuru.example.com",
-  "lock": {
-    "Locked": false,
-    "Reason": "",
-    "Owner": "",
-    "AcquireDate": "0001-01-01T00:00:00Z"
-  },
-  "name": "proj1-prod",
-  "owner": "webmaster@example.com",
-  "plan": {
-    "name": "medium",
-    "memory": 536870912,
-    "swap": 1073741824,
-    "cpushare": 1024,
-    "default": true,
-    "router": "hipache"
-  },
-  "platform": "python",
-  "pool": "mypool",
-  "repository": "git@example.com:proj1-prod.git",
-  "teamowner": "admin",
-  "teams": [
-    "admin",
-    "sysop"
-  ],
-  "units": [
-    {
-      "ID": "71281212acdf482318094afafafafafafabdbedf883197",
-      "Name": "",
-      "AppName": "proj1-prod",
-      "ProcessName": "web"
-    },
-    {
-      "ID": "71281184acdf482318193afafafafafafabdbedf183053",
-      "Name": "",
-      "AppName": "proj1-prod",
-      "ProcessName": "web"
-    },
-    {
-      "ID": "71282437acdf482318155afafafafafafabdbedf182839",
-      "Name": "",
-      "AppName": "proj1-prod",
-      "ProcessName": "web"
-    },
-    {
-      "ID": "71281271acdf482318180afafa0afafafabdbedf182883",
-      "Name": "",
-      "AppName": "proj1-prod",
-      "ProcessName": "web"
-    },
-    {
-      "ID": "71281341acdf482318123afafafafafafabdbedf182932",
-      "Name": "",
-      "AppName": "proj1-prod",
-      "ProcessName": "web"
-    }
-  ]
-}
-`
-
-const appInfo2 = `{
   "cname": ["proj3.dev.example.com"],
   "deploys": 35,
   "description": "my nice project",
@@ -295,7 +231,7 @@ const appInfo2 = `{
 }
 `
 
-const appInfo3 = `{
+const appInfo2 = `{
   "cname": ["proj3.example.com"],
   "deploys": 35,
   "description": "my nice project",
